Updating samples
The DU-Sense file will group examples from throughout the dictionary that have been reworked to try and improive the sense structure. The <form> group may not be stable, but the aim is to have it as close as can be. At this stage, I am not looking at citations or any structure lower than sense, re or etym.